Work around pygments lexer bug
This fix enables highlighting of the `flc_algorithm.i` file.

Pygments only looks for `r'##*[a-zA-Z_]\w*'`, but it should also look
for `r'##\s*\b'`. The pygments repo uses bitbucket and mercurial, and it
has a huge backlog, so realistically I'm not going to bother submitting
a patch upstream.